pyproject.toml
[tool.ruff.lint]
select = ["E", "F"]
ignore = ["E501", "E402"]

[tool.poetry]
name = "enviroflow-app"
version = "0.1.0"
description = "Make the business data sing"
authors = ["tofusoul <haoyen.shih@gmail.com>"]
readme = "README.md"
# packages = [{include = "enviroflow_app"}]

[tool.poetry.dependencies]
duckdb = "~1.2.0"
firebase-admin = "^6.1.0"
flask = "^2.3.2"
flask-oauthlib = "~0.9.6"
flask-session = "^0.8.0"
google-cloud-firestore = "^2.11.0"
google-cloud-storage = "^2.14.0"
gspread-pandas = "^3.2.2"
httpx = "^0.24.0"
jinja2 = "^3.1.3"
loguru = "^0.6.0"
mitosheet = "^0.1.532"
orjson = "^3.9.10"
palettable = "^3.3.3"
pandas = "^2.0.1"
pdfplumber = "^0.8.0"
pickleshare = "^0.7.5"
pillow = "^9.4.0"
plotly = "^5.14.0"
polars = { extras = ["pyarrow"], version = "^1.0.0" }
# prefect = "^2.14.5"
pyarrow = "^15.0.0"
pydantic = "^2.7.0"
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
python = "~3.10"
sqlalchemy = "^2.0.26"
streamlit = "^1.30.0"
tabulate = "^0.9.0"
typer = "^0.9.0"
rich = "^13.0.0"
validators = "^0.22.0"
watchdog = "~3.0.0"
xero-python = "^9.0.0"
xlsx2csv = "^0.8.2"
pygwalker = "^0.4.8.5"
matplotlib = "^3.10.0"
networkx = "^3.4.2"
minio = "^7.2.15"
# libsql-experimental = "^0.0.41"
# sqlmodel = "^0.0.24"
# sqlalchemy-libsql = "^0.1.0"
marimo = "^0.15.0"
basedpyright = "^1.31.4"
pylance = "^0.36.0"
toml = "^0.10.2"
sqlparse = "^0.5.3"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
mypy = "^1.2.0"
ipykernel = "^6.22.0"
ruff = "^0.11.7"
pdoc3 = "^0.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests",
]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.pyright]
venvPath = "."
venv = ".venv"
reportPrivateLocalImportUsage = "none"
reportUnusedCallResult = "none"
reportAny = "none"
reportExplicitAny = "none"
reportUnknownMemberType = "none"
reportUnnecessaryComparison = "none"
reportRedeclaration = "none"
